requirements.txt
google-adk   
google-genai
litellm
dotenv